requirements.txt
annotated-types
anyio
asgiref
boto3
botocore
certifi
charset-normalizer
click
fastapi
h11
idna
image
jmespath
pillow
pydantic
pydantic_core
python-dateutil
python-dotenv
python-multipart
requests
s3transfer
six
sniffio
sqlparse
starlette
typing_extensions
urllib3
uvicorn
